LPS-56580 Rename package to com.liferay.service.access.policy and rename SACPEntry to ServiceAccessPolicy
