+++ modules/apps/service-access-policy/service-access-policy-api/bnd.bnd
@@ -1,5 +1,5 @@
 Bundle-Name: Liferay Service Access Policy API
 Bundle-SymbolicName: com.liferay.service.access.policy.api
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.service.access.control.profile.*
+Export-Package: com.liferay.service.access.policy.*
 Include-Resource: classes
