+++ modules/apps/service-access-policy/service-access-policy-service/bnd.bnd
@@ -2,7 +2,7 @@ Bundle-Name: Liferay Service Access Policy Service
 Bundle-SymbolicName: com.liferay.service.access.policy.service
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.service.access.control.profile.service.permission
+	com.liferay.service.access.policy.service.permission
 Include-Resource: classes
 Liferay-Service: true
 Require-Capability:\
