+++ modules/apps/service-access-policy/service-access-policy-service/service.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0"?>
 <!DOCTYPE service-builder PUBLIC "-//Liferay//DTD Service Builder 7.0.0//EN" "http://www.liferay.com/dtd/liferay-service-builder_7_0_0.dtd">
 
-<service-builder package-path="com.liferay.service.access.control.profile">
-	<namespace>SACP</namespace>
-	<entity name="SACPEntry" uuid="true" local-service="true" remote-service="true">
+<service-builder package-path="com.liferay.service.access.policy">
+	<namespace>SAP</namespace>
+	<entity name="ServiceAccessPolicy" uuid="true" local-service="true" remote-service="true">
 
 		<!-- PK fields -->
 
-		<column name="sacpEntryId" type="long" primary="true" />
+		<column name="serviceAccessPolicyId" type="long" primary="true" />
 
 		<!-- Audit fields -->
 
@@ -20,7 +20,7 @@
 		<!-- Other fields -->
 
 		<column name="allowedServiceSignatures" type="String" />
-		<column name="defaultSACPEntry" type="boolean" />
+		<column name="defaultServiceAccessPolicy" type="boolean" />
 		<column name="name" type="String" />
 		<column name="title" type="String" localized="true" />
 
@@ -29,7 +29,7 @@
 		<finder name="CompanyId" return-type="Collection">
 			<finder-column name="companyId" />
 		</finder>
-		<finder name="C_N" return-type="SACPEntry">
+		<finder name="C_N" return-type="ServiceAccessPolicy">
 			<finder-column name="companyId" />
 			<finder-column name="name" case-sensitive="false" />
 		</finder>
@@ -40,9 +40,9 @@
 		<reference package-path="com.liferay.portal" entity="Role" />
 	</entity>
 	<exceptions>
-		<exception>DuplicateSACPEntryName</exception>
-		<exception>RequiredSACPEntry</exception>
-		<exception>SACPEntryName</exception>
-		<exception>SACPEntryTitle</exception>
+		<exception>DuplicateServiceAccessPolicyName</exception>
+		<exception>RequiredServiceAccessPolicy</exception>
+		<exception>ServiceAccessPolicyName</exception>
+		<exception>ServiceAccessPolicyTitle</exception>
 	</exceptions>
 </service-builder>
