+++ modules/apps/service-access-policy/service-access-policy-service/src/META-INF/portlet-model-hints.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 
 <model-hints>
-	<model name="com.liferay.service.access.control.profile.model.SACPEntry">
+	<model name="com.liferay.service.access.policy.model.ServiceAccessPolicy">
 		<field name="uuid" type="String" />
-		<field name="sacpEntryId" type="long" />
+		<field name="serviceAccessPolicyId" type="long" />
 		<field name="companyId" type="long" />
 		<field name="userId" type="long" />
 		<field name="userName" type="String" />
@@ -12,7 +12,7 @@
 		<field name="allowedServiceSignatures" type="String">
 			<hint-collection name="TEXTAREA" />
 		</field>
-		<field name="defaultSACPEntry" type="boolean" />
+		<field name="defaultServiceAccessPolicy" type="boolean" />
 		<field name="name" type="String" />
 		<field name="title" type="String" localized="true" />
 	</model>
