+++ modules/apps/service-access-policy/service-access-policy-service/src/META-INF/resource-actions/default.xml
@@ -3,28 +3,28 @@
 
 <resource-action-mapping>
 	<model-resource>
-		<model-name>com.liferay.service.access.control.profile</model-name>
+		<model-name>com.liferay.service.access.policy</model-name>
 		<portlet-ref>
-			<portlet-name>com_liferay_service_access_control_profile_web_portlet_SACPPortlet</portlet-name>
+			<portlet-name>com_liferay_service_access_policy_web_portlet_ServiceAccessPolicyPPortlet</portlet-name>
 		</portlet-ref>
 		<root>true</root>
 		<weight>1</weight>
 		<permissions>
 			<supports>
-				<action-key>ADD_SACP_ENTRY</action-key>
+				<action-key>ADD_SERVICE_ACCESS_POLICY</action-key>
 				<action-key>VIEW</action-key>
 			</supports>
 			<site-member-defaults />
 			<guest-defaults />
 			<guest-unsupported>
-				<action-key>ADD_SACP_ENTRY</action-key>
+				<action-key>ADD_SERVICE_ACCESS_POLICY</action-key>
 			</guest-unsupported>
 		</permissions>
 	</model-resource>
 	<model-resource>
-		<model-name>com.liferay.service.access.control.profile.model.SACPEntry</model-name>
+		<model-name>com.liferay.service.access.policy.model.ServiceAccessPolicy</model-name>
 		<portlet-ref>
-			<portlet-name>com_liferay_service_access_control_profile_web_portlet_SACPPortlet</portlet-name>
+			<portlet-name>com_liferay_service_access_policy_web_portlet_ServiceAccessPolicyPortlet</portlet-name>
 		</portlet-ref>
 		<weight>2</weight>
 		<permissions>
