--- modules/apps/service-access-policy/service-access-policy-service/src/com/liferay/service/access/control/profile/service/permission/SACPEntryPermission.java
@@ -1,86 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.service.access.control.profile.service.permission;
-
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.security.auth.PrincipalException;
-import com.liferay.portal.security.permission.BaseModelPermissionChecker;
-import com.liferay.portal.security.permission.PermissionChecker;
-import com.liferay.service.access.control.profile.model.SACPEntry;
-import com.liferay.service.access.control.profile.service.SACPEntryLocalServiceUtil;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Mika Koivisto
- */
-@Component(
-	property = {
-		"model.class.name=com.liferay.service.access.control.profile.model.SACPEntry"
-	}
-)
-public class SACPEntryPermission implements BaseModelPermissionChecker {
-
-	public static void check(
-			PermissionChecker permissionChecker, long sacpEntryId,
-			String actionId)
-		throws PortalException {
-
-		if (!contains(permissionChecker, sacpEntryId, actionId)) {
-			throw new PrincipalException.MustHavePermission(
-				permissionChecker, SACPEntry.class.getName(), sacpEntryId,
-				actionId);
-		}
-	}
-
-	public static void check(
-			PermissionChecker permissionChecker, SACPEntry sacpEntry,
-			String actionId)
-		throws PortalException {
-
-		if (!contains(permissionChecker, sacpEntry, actionId)) {
-			throw new PrincipalException.MustHavePermission(
-				permissionChecker, SACPEntry.class.getName(),
-				sacpEntry.getSacpEntryId(), actionId);
-		}
-	}
-
-	public static boolean contains(
-			PermissionChecker permissionChecker, long classPK, String actionId)
-		throws PortalException {
-
-		SACPEntry sacpEntry = SACPEntryLocalServiceUtil.getSACPEntry(classPK);
-
-		return contains(permissionChecker, sacpEntry, actionId);
-	}
-
-	public static boolean contains(
-		PermissionChecker permissionChecker, SACPEntry sacpEntry,
-		String actionId) {
-
-		return permissionChecker.hasPermission(
-			0, SACPEntry.class.getName(), sacpEntry.getSacpEntryId(), actionId);
-	}
-
-	@Override
-	public void checkBaseModel(
-			PermissionChecker permissionChecker, long groupId, long primaryKey,
-			String actionId)
-		throws PortalException {
-
-		check(permissionChecker, primaryKey, actionId);
-	}
-
-}
