+++ modules/apps/service-access-policy/service-access-policy-service/src/com/liferay/service/access/policy/ServiceAccessPolicyAccessControlPolicy.java
@@ -1,0 +1,277 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.service.access.policy;
+
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.security.access.control.AccessControlPolicy;
+import com.liferay.portal.kernel.security.access.control.AccessControlUtil;
+import com.liferay.portal.kernel.security.access.control.AccessControlled;
+import com.liferay.portal.kernel.security.access.control.BaseAccessControlPolicy;
+import com.liferay.portal.kernel.security.access.control.profile.ServiceAccessControlProfileThreadLocal;
+import com.liferay.portal.kernel.security.auth.verifier.AuthVerifierResult;
+import com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator;
+import com.liferay.portal.kernel.settings.SettingsException;
+import com.liferay.portal.kernel.settings.SettingsFactory;
+import com.liferay.portal.kernel.util.CharPool;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.security.auth.AccessControlContext;
+import com.liferay.portal.security.auth.CompanyThreadLocal;
+import com.liferay.service.access.policy.configuration.ServiceAccessPolicyConfiguration;
+import com.liferay.service.access.policy.constants.ServiceAccessPolicyConstants;
+import com.liferay.service.access.policy.model.ServiceAccessPolicy;
+import com.liferay.service.access.policy.service.ServiceAccessPolicyLocalService;
+
+import java.lang.reflect.Method;
+
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Mika Koivisto
+ */
+@Component(service = AccessControlPolicy.class)
+public class ServiceAccessPolicyAccessControlPolicy
+	extends BaseAccessControlPolicy {
+
+	@Override
+	public void onServiceRemoteAccess(
+			Method method, Object[] arguments,
+			AccessControlled accessControlled)
+		throws SecurityException {
+
+		List<String> serviceAccessControlProfileNames =
+			ServiceAccessControlProfileThreadLocal.
+				getActiveServiceAccessControlProfileNames();
+
+		ServiceAccessPolicyConfiguration serviceAccessPolicyConfiguration =
+			null;
+
+		try {
+			serviceAccessPolicyConfiguration = _settingsFactory.getSettings(
+				ServiceAccessPolicyConfiguration.class,
+				new CompanyServiceSettingsLocator(
+					CompanyThreadLocal.getCompanyId(),
+					ServiceAccessPolicyConstants.SERVICE_NAME));
+		}
+		catch (SettingsException se) {
+			throw new SecurityException(
+				"Unable to determine default service access control profile",
+				se);
+		}
+
+		if (serviceAccessPolicyConfiguration.requireDefaultServiceAccessPolicy(
+				) ||
+			(serviceAccessControlProfileNames == null)) {
+
+			if (serviceAccessControlProfileNames == null) {
+				serviceAccessControlProfileNames = new ArrayList<>();
+
+				ServiceAccessControlProfileThreadLocal.
+					setActiveServiceAccessControlProfileNames(
+						serviceAccessControlProfileNames);
+			}
+
+			boolean passwordBasedAuthentication = false;
+
+			AccessControlContext accessControlContext =
+				AccessControlUtil.getAccessControlContext();
+
+			if (accessControlContext != null) {
+				AuthVerifierResult authVerifierResult =
+					accessControlContext.getAuthVerifierResult();
+
+				if (authVerifierResult != null) {
+					passwordBasedAuthentication =
+						authVerifierResult.isPasswordBasedAuthentication();
+				}
+			}
+
+			if (passwordBasedAuthentication) {
+				serviceAccessControlProfileNames.add(
+					serviceAccessPolicyConfiguration.
+						defaultUserServiceAccessPolicyName());
+			}
+			else {
+				serviceAccessControlProfileNames.add(
+					serviceAccessPolicyConfiguration.
+						defaultApplicationServiceAccessPolicyName());
+			}
+		}
+
+		long companyId = CompanyThreadLocal.getCompanyId();
+
+		Set<String> allowedServiceSignatures = new HashSet<>();
+
+		for (String name : serviceAccessControlProfileNames) {
+			try {
+				ServiceAccessPolicy serviceAccessPolicy =
+					_serviceAccessPolicyLocalService.getServiceAccessPolicy(
+						companyId, name);
+
+				allowedServiceSignatures.addAll(
+					serviceAccessPolicy.getAllowedServiceSignaturesList());
+			}
+			catch (PortalException pe) {
+				throw new SecurityException(pe);
+			}
+		}
+
+		if (allowedServiceSignatures.contains(StringPool.STAR)) {
+			return;
+		}
+
+		Class<?> clazz = method.getDeclaringClass();
+
+		String className = clazz.getName();
+
+		if (allowedServiceSignatures.contains(className)) {
+			return;
+		}
+
+		String methodName = method.getName();
+
+		String classNameAndMethodName = className.concat(
+			StringPool.POUND).concat(methodName);
+
+		if (allowedServiceSignatures.contains(classNameAndMethodName)) {
+			return;
+		}
+
+		for (String allowedService : allowedServiceSignatures) {
+			if (matches(className, methodName, allowedService)) {
+				return;
+			}
+		}
+
+		throw new SecurityException(
+			"Access denied to " + classNameAndMethodName);
+	}
+
+	protected boolean matches(
+		String className, String methodName, String allowedServiceSignature) {
+
+		String allowedClassName = null;
+		String allowedMethodName = null;
+
+		int index = allowedServiceSignature.indexOf(CharPool.POUND);
+
+		if (index > -1) {
+			allowedClassName = allowedServiceSignature.substring(0, index);
+			allowedMethodName = allowedServiceSignature.substring(index + 1);
+		}
+		else {
+			allowedClassName = allowedServiceSignature;
+		}
+
+		boolean wildcardMatchClass = false;
+
+		if (Validator.isNotNull(allowedClassName) &&
+			allowedClassName.endsWith(StringPool.STAR)) {
+
+			allowedClassName = allowedClassName.substring(
+				0, allowedClassName.length() - 1);
+			wildcardMatchClass = true;
+		}
+
+		boolean wildcardMatchMethod = false;
+
+		if (Validator.isNotNull(allowedMethodName) &&
+			allowedMethodName.endsWith(StringPool.STAR)) {
+
+			allowedMethodName = allowedMethodName.substring(
+				0, allowedMethodName.length() - 1);
+			wildcardMatchMethod = true;
+		}
+
+		if (Validator.isNotNull(allowedClassName) &&
+			Validator.isNotNull(allowedMethodName)) {
+
+			if (wildcardMatchClass && !className.startsWith(allowedClassName)) {
+				return false;
+			}
+			else if (!wildcardMatchClass &&
+					 !className.equals(allowedClassName)) {
+
+				return false;
+			}
+
+			if (wildcardMatchMethod &&
+				!methodName.startsWith(allowedMethodName)) {
+
+				return false;
+			}
+			else if (!wildcardMatchMethod &&
+					 !methodName.equals(allowedMethodName)) {
+
+				return false;
+			}
+
+			return true;
+		}
+		else if (Validator.isNotNull(allowedClassName)) {
+			if (wildcardMatchClass && !className.startsWith(allowedClassName)) {
+				return false;
+			}
+			else if (!wildcardMatchClass &&
+					 !className.equals(allowedClassName)) {
+
+				return false;
+			}
+
+			return true;
+		}
+		else if (Validator.isNotNull(allowedMethodName)) {
+			if (wildcardMatchMethod &&
+				!methodName.startsWith(allowedMethodName)) {
+
+				return false;
+			}
+			else if (!wildcardMatchMethod &&
+					 !methodName.equals(allowedMethodName)) {
+
+				return false;
+			}
+
+			return true;
+		}
+		else if (wildcardMatchClass && Validator.isNull(allowedClassName)) {
+			return true;
+		}
+
+		return false;
+	}
+
+	@Reference(unbind = "-")
+	protected void setServiceAccessPolicyLocalService(
+		ServiceAccessPolicyLocalService serviceAccessPolicyLocalService) {
+
+		_serviceAccessPolicyLocalService = serviceAccessPolicyLocalService;
+	}
+
+	@Reference
+	protected void setSettingsFactory(SettingsFactory settingsFactory) {
+		_settingsFactory = settingsFactory;
+	}
+
+	private ServiceAccessPolicyLocalService _serviceAccessPolicyLocalService;
+	private volatile SettingsFactory _settingsFactory;
+
+}
