+++ modules/apps/service-access-policy/service-access-policy-service/src/portlet.properties
@@ -6,4 +6,4 @@ include-and-override=portlet-ext.properties
 #
 resource.actions.configs=META-INF/resource-actions/default.xml
 
-service.configurator.portlet.ids=com_liferay_service_access_control_profile_web_portlet_SACPPortlet
+service.configurator.portlet.ids=com_liferay_service_access_policy_web_portlet_ServiceAccessPolicyPortlet
