+++ modules/apps/service-access-policy/service-access-policy-web/src/META-INF/resources/init.jsp
@@ -32,15 +32,14 @@ page import="com.liferay.portal.kernel.util.WebKeys" %><%@
 page import="com.liferay.portal.security.permission.ActionKeys" %><%@
 page import="com.liferay.portal.service.permission.PortletPermissionUtil" %><%@
 page import="com.liferay.portal.util.PortalUtil" %><%@
-page import="com.liferay.service.access.control.profile.constants.SACPActionKeys" %><%@
-page import="com.liferay.service.access.control.profile.constants.SACPConstants" %><%@
-page import="com.liferay.service.access.control.profile.exception.DuplicateSACPEntryNameException" %><%@
-page import="com.liferay.service.access.control.profile.exception.SACPEntryNameException" %><%@
-page import="com.liferay.service.access.control.profile.exception.SACPEntryTitleException" %><%@
-page import="com.liferay.service.access.control.profile.model.SACPEntry" %><%@
-page import="com.liferay.service.access.control.profile.model.SACPEntryConstants" %><%@
-page import="com.liferay.service.access.control.profile.service.SACPEntryServiceUtil" %><%@
-page import="com.liferay.service.access.control.profile.service.permission.SACPEntryPermission" %><%@
+page import="com.liferay.service.access.policy.constants.ServiceAccessPolicyActionKeys" %><%@
+page import="com.liferay.service.access.policy.constants.ServiceAccessPolicyConstants" %><%@
+page import="com.liferay.service.access.policy.exception.DuplicateServiceAccessPolicyNameException" %><%@
+page import="com.liferay.service.access.policy.exception.ServiceAccessPolicyNameException" %><%@
+page import="com.liferay.service.access.policy.exception.ServiceAccessPolicyTitleException" %><%@
+page import="com.liferay.service.access.policy.model.ServiceAccessPolicy" %><%@
+page import="com.liferay.service.access.policy.service.ServiceAccessPolicyServiceUtil" %><%@
+page import="com.liferay.service.access.policy.service.permission.ServiceAccessPolicyPermission" %><%@
 page import="com.liferay.taglib.search.ResultRow" %>
 
 <%@ page import="javax.portlet.PortletURL" %>
